Issue #217 -- Continued refactor of version id.
